+++ src/HotChocolate/Core/test/Types.Tests/Types/Scalars/ScalarsTests.cs
@@ -5,7 +5,7 @@ public class ScalarsTests
     [InlineData(Foo.Bar, ValueKind.Enum)]
     [InlineData("foo", ValueKind.String)]
     [InlineData((short)1, ValueKind.Integer)]
-    [InlineData((int)1, ValueKind.Integer)]
+    [InlineData(1, ValueKind.Integer)]
     [InlineData((long)1, ValueKind.Integer)]
     [InlineData((ushort)1, ValueKind.Integer)]
     [InlineData((uint)1, ValueKind.Integer)]
@@ -27,7 +27,7 @@ public class ScalarsTests
 
     [InlineData(Foo.Bar, ValueKind.Enum)]
     [InlineData((short)1, ValueKind.Integer)]
-    [InlineData((int)1, ValueKind.Integer)]
+    [InlineData(1, ValueKind.Integer)]
     [InlineData((long)1, ValueKind.Integer)]
     [InlineData((ushort)1, ValueKind.Integer)]
     [InlineData((uint)1, ValueKind.Integer)]
